Transitioning to modern cmake
The modern cmake approach is to be more target centric and fine
grained than the cmake written here.  The end goal is to merge the
several libraries that are written for host, and device, and then
linked together into one library with CMake, using separable
compilation.

First, modernize the CMakeLists.txt as-is, and then add extras.