Continues work on CMake Modernization
Most of the work in this commit relates to updating the CMakeLists.  There are
a very limited set of code changes, relating to missing header files for stdexcept.

One addition is the use of FetchContent over git submodules.  This is experimental
at the moment, as Eigen has very many flags that need adding to build 'correctly'.
An option was added for the use of an external eigen3, which I suspect most people
already have, or can get, without having an additional version living in thirdparty.