Removes redundant example code
The two sets up examples (numerical and analytical) are highly
redundant, making maintenance more difficult for future purposes.
Since the difference was only the data type, a single source
now serves each, with a pre-processor macro definition controlling
the typedef in each file.

Other updates include a a re-written CMakeLists.txt for the examples
with a few TODO's scattered throughout the code.